setup automatic build sys for travis
