+++ .travis.yml
@@ -1,0 +1,3 @@
+language: c
+
+script: make
